Wheel of Fortune
The game is played with a special wheel that determines the dollar value of the called letters. The wheel is split into 10 segments as follows: 10, 20, 30, 40, 100, 200, 300, 400, Bankrupt, and Bankrupt.
The computer randomly determines a word and then a player tries to guess it by suggesting letters. The word should be an animal in English such as dog, cat, chicken, penguin, etc.